move models to new file
