+++ src/app/shared/models/grid/grid-events.model.ts
@@ -1,0 +1,14 @@
+export type BooleanChange<T> = {
+  value: boolean;
+  item: T;
+};
+
+export type RowReorderingEvent<T> = {
+  from: Indexed<T>;
+  to: Indexed<T>;
+};
+
+export type Indexed<T> = {
+  item: T;
+  index: number;
+};
